move uploading logic to uploader
